Move directory_hash_id into galaxy/util.py. Use util.directory_hash_id throughout.
