+++ lib/galaxy/webapps/tool_shed/model/__init__.py
@@ -507,17 +507,3 @@ def sort_by_attr( seq, attr ):
     intermed = map( None, map( getattr, seq, ( attr, ) * len( seq ) ), xrange( len( seq ) ), seq )
     intermed.sort()
     return map( operator.getitem, intermed, ( -1, ) * len( intermed ) )
-
-
-def directory_hash_id( id ):
-    s = str( id )
-    l = len( s )
-    # Shortcut -- ids 0-999 go under ../000/
-    if l < 4:
-        return [ "000" ]
-    # Pad with zeros until a multiple of three
-    padded = ( ( ( 3 - len( s ) ) % 3 ) * "0" ) + s
-    # Drop the last three digits -- 1000 files per directory
-    padded = padded[:-3]
-    # Break into chunks of three
-    return [ padded[i * 3:(i + 1) * 3] for i in range( len( padded ) // 3 ) ]
